Remove both length warning logs
It's common for the enclosure length attribute to not match the
content-length header, so this is just noise.

Checking the number of bytes written against the HTTP content-length
header is potentially more meaningful, as you want the written file to
correspond exactly to the bytes received and not, eg, be truncated.

However, the content-length header may not be provided when the file is
streamed out (and so the size is not known up front). See MDN. In fact,
since we previously unconditionally `unwrap`ed the header, it's possible
that this would cause a crash.